refactor(perf): split /insights + /rituals into server shell + motion island
21st + 22nd in the pattern series.
