+++ components/rituals/rituals-data.ts
@@ -1,0 +1,26 @@
+export const ritualsMeta = [
+  {
+    title: 'Morning Prime',
+    time: '6:00 — 7:30 AM',
+    description:
+      'Set the creative direction before the world starts pulling. Review overnight ideas, pick the one thread worth pulling, and write the first 200 words.',
+  },
+  {
+    title: 'Deep Build',
+    time: '9:00 AM — 12:30 PM',
+    description:
+      '3.5 hours of uninterrupted building. No Slack, no email, no meetings. This is where the ACOS gets developed, articles get written, and products get shipped.',
+  },
+  {
+    title: 'Studio Session',
+    time: '9:00 PM — 12:00 AM',
+    description:
+      'The night studio. This is where 12,000+ songs came from — late-night sessions with Suno, experimenting across genres, pushing what AI music can do.',
+  },
+  {
+    title: 'Evening Review',
+    time: '11:30 PM — 12:00 AM',
+    description:
+      'Clear the day\'s residue. Review what shipped, capture ideas that surfaced during the session, and set tomorrow\'s intention.',
+  },
+]
